修正倒计时错误

+++ src/main/java/com/group7/sanSongMall/entity/orders.java
@@ -39,7 +39,7 @@ public class orders {
 
     private Integer state;
 
-    private Integer delay;
+    private Long delay;
 
     private Timestamp orderTime;
